correcciones

+++ Entities/MoviesList.cs
@@ -2,13 +2,13 @@
 {
     public class MoviesList
     {
-        public long page { get; set; }
+        public int page { get; set; }
 
         public List<Movie> results { get; set; }
 
-        public long total_pages { get; set; }
+        public int total_pages { get; set; }
 
-        public long total_results { get; set; }
+        public int total_results { get; set; }
 
         public MoviesList()
         {
